requirements.txt
aiogram>=3.13,<4
asyncpg>=0.30,<1
python-dotenv>=1.0,<2
